Update ec2_launch_wizard_security_group_prohibited.py
This proposed change updates the code to handle deleted resources. Currently, if an ENI is deleted and triggers a custom AWS Config rule, the existing code will not provide a return code to remove that deleted resource. This update adds in that logic and also subsequently modifies the Security Group checks to be performed if the resource is **not** deleted.